Fix office not working when activeAirline.stats is undefined
when activeAirline is updated without extendedInfo=true, which is the case when refreshPanels is called, the resulting object does not have `stats`; in the office menu, the game would try to update with non-existent data and error out. This stops the game from trying to update and to just move on.